Task 0

+++ 0x12-singly_linked_lists/0-print_list.c
@@ -8,14 +8,14 @@
  */
 size_t print_list(const list_t *h)
 {
-	int count = 0;
+       size_t count = 0;
 
 	while (h)
 	{
 		if (!h->str)
 			printf("[0] (nil)\n");
 		else
-			printf("[%d] %s\n", h->len, h->str);
+			printf("[%u] %s\n", h->len, h->str);
 		h = h->next;
 		count++;
 	}
